# IGNITE-25586 Sql. Partition awareness. Performance tests