--- a/modules/client/src/main/java/org/apache/ignite/internal/client/sql/ClientSql.java
+++ b/modules/client/src/main/java/org/apache/ignite/internal/client/sql/ClientSql.java
@@ -291,7 +291,7 @@ public class ClientSql implements IgniteSql {
 
         PartitionMappingProvider mappingProvider = mappingProviderCache.getIfPresent(new PaCacheKey(statement));
 
-        PartitionMapping mapping = mappingProvider != null 
+        PartitionMapping mapping = mappingProvider != null
                 ? mappingProvider.get(arguments)
                 : null;