--- a/modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/BulkLoadBenchmark.java
+++ b/modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/BulkLoadBenchmark.java
@@ -116,9 +116,13 @@ public class BulkLoadBenchmark extends AbstractMultiNodeBenchmark {
         public void setUp() {
             String queryStr = createInsertStatement();
 
-            client = IgniteClient.builder()
-                    .addresses("127.0.0.1:10800", "127.0.0.1:10801", "127.0.0.1:10802")
-                    .build();
+            client = IgniteClient.builder().addresses("127.0.0.1:10800").build();
+
+            String[] clientAddrs = getServerEndpoints(client);
+
+            client.close();
+
+            client = IgniteClient.builder().addresses(clientAddrs).build();
 
             sql = client.sql();
 
@@ -130,8 +134,7 @@ public class BulkLoadBenchmark extends AbstractMultiNodeBenchmark {
          */
         @TearDown
         public void tearDown() throws Exception {
-            // statement.close() throws `UnsupportedOperationException("Not implemented yet.")`, that's why it's commented.
-            closeAll(/* statement, */ client);
+            closeAll(client);
         }
 
         void upload(int count, int batch) {
@@ -171,9 +174,13 @@ public class BulkLoadBenchmark extends AbstractMultiNodeBenchmark {
                 tuple.set("field" + i, FIELD_VAL);
             }
 
-            client = IgniteClient.builder()
-                    .addresses("127.0.0.1:10800", "127.0.0.1:10801", "127.0.0.1:10802")
-                    .build();
+            client = IgniteClient.builder().addresses("127.0.0.1:10800").build();
+
+            String[] clientAddrs = getServerEndpoints(client);
+
+            client.close();
+
+            client = IgniteClient.builder().addresses(clientAddrs).build();
 
             kvView = client.tables().table(TABLE_NAME).keyValueView();
         }